add double click for connections
